+++ test/utils_selection_tests.py
@@ -53,6 +53,9 @@ class MockPool(object):
     def close(self):
         pass
 
+    def update_is_writable(self, is_writable):
+        pass
+
     def remove_stale_sockets(self):
         pass